Clean Up Part 1
Remove Irrelevant Methods
Fix Warning
Added BUGS And TODO LIST

+++ src/client/MapleCharacter.java
@@ -31,8 +31,6 @@ import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.stream.Collectors;
 
-import org.apache.mina.common.WriteFuture;
-
 import client.MapleTrait.MapleTraitType;
 import client.inventory.Equip;
 import client.inventory.Item;
@@ -50,7 +48,6 @@ import client.inventory.MaplePotionPot;
 import client.inventory.MapleRing;
 import constants.GameConstants;
 import constants.MapConstants;
-import constants.ServerConstants;
 import handling.channel.ChannelServer;
 import handling.channel.handler.AttackInfo;
 import handling.channel.handler.PlayerHandler;
@@ -3713,7 +3710,7 @@ public class MapleCharacter extends AnimatedMapleMapObject implements Serializab
                 if (map.isTown() || FieldLimitType.RegularExpLoss.check(map.getFieldLimit())) {
                     diepercentage = 0.01f;
                 } else {
-                    diepercentage = (float) (0.1f - ((traits.get(MapleTraitType.charisma).getLevel() / 20) / 100f) - (stats.expLossReduceR / 100f));
+                    diepercentage = 0.1f - ((traits.get(MapleTraitType.charisma).getLevel() / 20) / 100f) - (stats.expLossReduceR / 100f);
                 }
                 long v10 = (exp - (long) ((double) expforlevel * diepercentage));
                 if (v10 < 0) {
@@ -5472,7 +5469,7 @@ public class MapleCharacter extends AnimatedMapleMapObject implements Serializab
         if (guildid <= 0) {
             return;
         }
-        mgc.setLevel((short) level);
+        mgc.setLevel(level);
         mgc.setJobId(job);
         World.Guild.memberLevelJobUpdate(mgc);
     }
@@ -6715,18 +6712,18 @@ public class MapleCharacter extends AnimatedMapleMapObject implements Serializab
         int total = stats.getStr() + stats.getDex() + stats.getLuk() + stats.getInt() + getRemainingAp();
 
         total -= str;
-        stats.str = (int) str;
+        stats.str = str;
 
         total -= dex;
-        stats.dex = (int) dex;
+        stats.dex = dex;
 
         total -= int_;
-        stats.int_ = (int) int_;
+        stats.int_ = int_;
 
         total -= luk;
-        stats.luk = (int) luk;
+        stats.luk = luk;
 
-        setRemainingAp((int) total);
+        setRemainingAp(total);
         stats.recalcLocalStats(this);
         stat.put(MapleStat.STR, (long) str);
         stat.put(MapleStat.DEX, (long) dex);
@@ -6858,7 +6855,7 @@ public class MapleCharacter extends AnimatedMapleMapObject implements Serializab
     public void expandInventory(byte type, int amount) {
         final MapleInventory inv = getInventory(MapleInventoryType.getByType(type));
         inv.addSlot((byte) amount);
-        client.getSession().write(InventoryPacket.getSlotUpdate(type, (byte) inv.getSlotLimit()));
+        client.getSession().write(InventoryPacket.getSlotUpdate(type, inv.getSlotLimit()));
     }
 
     public boolean allowedToTarget(MapleCharacter other) {
@@ -7734,7 +7731,7 @@ public class MapleCharacter extends AnimatedMapleMapObject implements Serializab
                         } else {
                             MapleCharacter chr = (MapleCharacter) attacke;
                             chr.addHP(-divineShield.getDamage());
-                            attack.add((int) divineShield.getDamage());
+                            attack.add(divineShield.getDamage());
                         }
                     }
                 }
@@ -7751,7 +7748,7 @@ public class MapleCharacter extends AnimatedMapleMapObject implements Serializab
                         } else {
                             final MapleCharacter attacker = (MapleCharacter) attacke;
                             attacker.addHP(-divineShield.getDamage());
-                            attack.add((int) divineShield.getDamage());
+                            attack.add(divineShield.getDamage());
                         }
                     }
                 }
@@ -7771,7 +7768,7 @@ public class MapleCharacter extends AnimatedMapleMapObject implements Serializab
                         } else {
                             final MapleCharacter attacker = (MapleCharacter) attacke;
                             attacker.addHP(-divineShield.getDamage());
-                            attack.add((int) divineShield.getDamage());
+                            attack.add(divineShield.getDamage());
                         }
                     }
                 }
@@ -7865,7 +7862,7 @@ public class MapleCharacter extends AnimatedMapleMapObject implements Serializab
             addHP(((int) Math.min(maxhp, Math.min(((int) ((double) totDamage * (double) getStatForBuff(MapleBuffStat.ComboDrain).getX() / 100.0)), stats.getMaxHp() / 2))));
         }
         if (getBuffSource(MapleBuffStat.ComboDrain) == 23101003) {
-            addMP(((int) Math.min(maxmp, Math.min(((int) ((double) totDamage * (double) getStatForBuff(MapleBuffStat.ComboDrain).getX() / 100.0)), stats.getMaxMp() / 2))));
+            addMP((Math.min(maxmp, Math.min(((int) ((double) totDamage * (double) getStatForBuff(MapleBuffStat.ComboDrain).getX() / 100.0)), stats.getMaxMp() / 2))));
         }
         if (getBuffedValue(MapleBuffStat.REAPER) != null && getBuffedValue(MapleBuffStat.SUMMON) == null && getSummonsSize() < 4 && canSummon()) {
             final MapleStatEffect eff = getStatForBuff(MapleBuffStat.REAPER);
