+++ src/client/SkillFactory.java
@@ -204,7 +204,7 @@ public class SkillFactory {
         if (skill == null) {
             return -1;
         }
-        return (skill.isTimeLimited() ? (System.currentTimeMillis() + (long) (30L * 24L * 60L * 60L * 1000L)) : -1);
+        return (skill.isTimeLimited() ? (System.currentTimeMillis() + 30L * 24L * 60L * 60L * 1000L) : -1);
     }
 
     public static CraftingEntry getCraft(final int id) {
