+++ src/server/maps/MapScriptMethods.java
@@ -30,7 +30,6 @@ import client.MapleClient;
 import client.Skill;
 import client.SkillEntry;
 import client.SkillFactory;
-import constants.GameConstants;
 import script.event.EventManager;
 import script.npc.NPCScriptManager;
 import server.MapleItemInformationProvider;
@@ -44,7 +43,6 @@ import server.maps.MapleNodes.DirectionInfo;
 import server.quest.MapleQuest;
 import server.quest.MapleQuestStatus;
 import server.quest.MapleQuest.MedalQuest;
-import tools.FileoutputUtil;
 import tools.Randomizer;
 import tools.packet.CField;
 import tools.packet.CField.EffectPacket;
@@ -454,7 +452,7 @@ public class MapScriptMethods {
                     c.getPlayer().getMap().startMapEffect("Don't forget that you have to clear it within the time limit! Take down the monster and head to the next floor!", 5120024);
                 }
                 int temp = (c.getPlayer().getMapId() - 925000000) / 100;
-                int stage = (int) (temp - ((temp / 100) * 100));
+                int stage = temp - ((temp / 100) * 100);
 //                String lol = c.getPlayer().getInfoQuest((int)7214);
 //                System.err.println("ol " + lol);
 //                int ad = Integer.parseInt(lol);
